+++ rspec-core/spec/rspec/core/dsl_spec.rb
@@ -1,17 +1,25 @@
 require 'spec_helper'
 
 main = self
-describe "The describe method" do
-  it 'is available on the main object' do
-    main.should respond_to(:describe)
-  end
 
-  it 'is available on modules (so example groups can be nested inside them)' do
-    Module.new.should respond_to(:describe)
-  end
+describe "The RSpec DSL" do
+  methods = [
+    :describe,
+    :share_examples_for,
+    :shared_examples_for,
+    :shared_examples,
+    :shared_context,
+    :share_as
+  ]
 
-  it 'is not available on other types of objects' do
-    Object.new.should_not respond_to(:describe)
+  methods.each do |method_name|
+    describe "##{method_name}" do
+      it "is not added to every object in the system" do
+        main.should respond_to(method_name)
+        Module.new.should respond_to(method_name)
+        Object.new.should_not respond_to(method_name)
+      end
+    end
   end
 end